Half star rating active, forcing sync, trimmed excess files

+++ models/song.js
@@ -25,7 +25,7 @@ module.exports = function(sequelize, DataTypes) {
       },
     },
     pilot_rating: {
-      type: DataTypes.INTEGER(1),
+      type: DataTypes.DECIMAL(2, 1),
       default: null,
       validate: {
         max: 5,
@@ -33,7 +33,7 @@ module.exports = function(sequelize, DataTypes) {
       },
     },
     copilot_rating: {
-      type: DataTypes.INTEGER(1),
+      type: DataTypes.DECIMAL(2, 1),
       default: null,
       validate: {
         max: 5,
@@ -41,7 +41,7 @@ module.exports = function(sequelize, DataTypes) {
       },
     },
     avg_rating: {
-      type: DataTypes.DECIMAL(2, 1),
+      type: DataTypes.DECIMAL(3, 2),
       default: null,
       validate: {
         max: 5,
